refactor package name for consistency
