+++ app/src/main/java/com/canli/oya/traininventoryroom/di/AppComponent.kt
@@ -9,8 +9,8 @@ import com.canli.oya.traininventoryroom.ui.brands.AddBrandFragment
 import com.canli.oya.traininventoryroom.ui.brands.BrandListFragment
 import com.canli.oya.traininventoryroom.ui.categories.AddCategoryFragment
 import com.canli.oya.traininventoryroom.ui.categories.CategoryListFragment
+import com.canli.oya.traininventoryroom.ui.filter.FilterTrainFragment
 import com.canli.oya.traininventoryroom.ui.main.MainActivity
-import com.canli.oya.traininventoryroom.ui.searchtrain.FilterTrainFragment
 import com.canli.oya.traininventoryroom.ui.trains.TrainDetailsFragment
 import com.canli.oya.traininventoryroom.ui.trains.TrainListFragment
 import dagger.Component
