+++ app/src/main/java/com/canli/oya/traininventoryroom/ui/filter/FilterBySpinnerAdapter.kt
@@ -1,4 +1,4 @@
-package com.canli.oya.traininventoryroom.ui.searchtrain
+package com.canli.oya.traininventoryroom.ui.filter
 
 import android.content.Context
 import android.widget.ArrayAdapter
