+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
     app:startDestination="@id/trainListFragment">
 
@@ -80,7 +79,7 @@
     </dialog>
     <fragment
         android:id="@+id/filterTrainFragment"
-        android:name="com.canli.oya.traininventoryroom.ui.searchtrain.FilterTrainFragment"
+        android:name="com.canli.oya.traininventoryroom.ui.filter.FilterTrainFragment"
         android:label="@string/search_trains" >
         <argument
             android:name="intentRequestCode"
